refactor(widgets): restructure common widgets into category-based folders for better organization
