+++ lib/common/widgets/helpers/cursor_manager.dart
