+++ lib/desktop/features/action_library/presentation/widgets/forms/single_field_binding_action_form.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:streamkeys/common/widgets/field_label.dart';
+import 'package:streamkeys/common/widgets/forms/field_label.dart';
 import 'package:streamkeys/core/constants/spacing.dart';
 
 class SingleFieldBindingActionForm extends StatefulWidget {
