+++ lib/desktop/features/hidmacros/presentation/widgets/select_keyboard_type.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:streamkeys/common/widgets/selectable_tile_list.dart';
+import 'package:streamkeys/common/widgets/tiles/selectable_tile_list.dart';
 import 'package:streamkeys/desktop/features/hidmacros/bloc/hidmacros_bloc.dart';
 import 'package:streamkeys/desktop/features/key_grid_area/data/models/keyboard_type.dart';
