+++ lib/desktop/features/key_bindings/presentation/widgets/key_color_field.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:streamkeys/common/widgets/color_picker/custom_color_picker.dart';
-import 'package:streamkeys/common/widgets/field_label.dart';
+import 'package:streamkeys/common/widgets/forms/field_label.dart';
 import 'package:streamkeys/core/constants/spacing.dart';
 
 class KeyColorField extends StatefulWidget {
