+++ lib/desktop/features/key_bindings/presentation/widgets/key_image_field.dart
@@ -1,8 +1,8 @@
 import 'package:dotted_border/dotted_border.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
-import 'package:streamkeys/common/widgets/field_label.dart';
-import 'package:streamkeys/common/widgets/image_preview.dart';
+import 'package:streamkeys/common/widgets/forms/field_label.dart';
+import 'package:streamkeys/common/widgets/image/image_preview.dart';
 import 'package:streamkeys/core/constants/colors.dart';
 import 'package:streamkeys/core/constants/spacing.dart';
 import 'package:streamkeys/core/constants/typography.dart';
