+++ lib/desktop/features/key_bindings/presentation/widgets/key_name_field.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:streamkeys/common/widgets/field_label.dart';
+import 'package:streamkeys/common/widgets/forms/field_label.dart';
 import 'package:streamkeys/core/constants/spacing.dart';
 import 'package:streamkeys/core/constants/typography.dart';
