+++ lib/desktop/features/key_grid_area/presentation/widgets/grid/grid_area_stack.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:streamkeys/common/widgets/custom_dropdown_button.dart';
+import 'package:streamkeys/common/widgets/buttons/custom_dropdown_button.dart';
 import 'package:streamkeys/core/constants/spacing.dart';
 import 'package:streamkeys/desktop/features/key_grid_area/data/models/grid_template.dart';
 import 'package:streamkeys/desktop/features/key_grid_area/presentation/widgets/grid/grid_area_wrapper.dart';
