+++ lib/desktop/features/key_grid_area/presentation/widgets/key_grid_area.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:streamkeys/common/widgets/draggable_within_parent.dart';
+import 'package:streamkeys/common/widgets/helpers/draggable_within_parent.dart';
 import 'package:streamkeys/core/constants/colors.dart';
 import 'package:streamkeys/core/constants/spacing.dart';
