+++ lib/desktop/features/key_grid_area/presentation/widgets/keyboard/keyboard_area_stack.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:streamkeys/common/widgets/custom_dropdown_button.dart';
+import 'package:streamkeys/common/widgets/buttons/custom_dropdown_button.dart';
 import 'package:streamkeys/core/constants/spacing.dart';
 import 'package:streamkeys/desktop/features/hidmacros/bloc/hidmacros_bloc.dart';
 import 'package:streamkeys/desktop/features/hidmacros/presentation/widgets/select_keyboard.dart';
