+++ lib/desktop/features/key_grid_area/presentation/widgets/keyboard/keyboard_blocks/main_keys_block.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:streamkeys/common/widgets/for.dart';
+import 'package:streamkeys/common/widgets/helpers/for.dart';
 import 'package:streamkeys/core/constants/spacing.dart';
 import 'package:streamkeys/desktop/features/key_grid_area/data/models/keyboard_key_data.dart';
 import 'package:streamkeys/desktop/features/key_grid_area/presentation/widgets/keyboard/keyboard_blocks/base_keys_block.dart';
