+++ lib/desktop/features/settings/presentation/widgets/http_server_config/server_password_field.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:streamkeys/common/widgets/copy_icon_button.dart';
-import 'package:streamkeys/common/widgets/small_icon_button.dart';
+import 'package:streamkeys/common/widgets/buttons/copy_icon_button.dart';
+import 'package:streamkeys/common/widgets/buttons/small_icon_button.dart';
 import 'package:streamkeys/core/constants/spacing.dart';
 import 'package:streamkeys/service_locator.dart';
