+++ test/dashboard/page_tab_mock.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:streamkeys/common/widgets/page_tab.dart';
+import 'package:streamkeys/common/widgets/tabs/page_tab.dart';
 
 class PageTabMock extends StatelessWidget with PageTab {
   @override
